Remove checkMapping function
